requirements.txt
fastapi
uvicorn
sqlmodel
pydantic>=2.0
httpx
python-dotenv
psycopg2-binary
# alembic